Cleaning.R Overhaul
After noticing the discrepancies inherit in my cleaning/ importing the date-times, I revamped how the data are loaded. I relied on the dmy_hr function from the lubridate package to bring the dates into R, instead of using my own functions. I have aligned the results with the CSV files from DPD, and it looks like we are good to go.